+++ src/HearThis/UI/AdministrativeSettings.Designer.cs
@@ -36,8 +36,8 @@ namespace HearThis.UI
             System.Windows.Forms.Button _btnCancel;
             System.Windows.Forms.Label _lblRecordByVerseRangeStart;
             System.Windows.Forms.Label _lblRecordByVerseRangeEnd;
-            System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(AdministrativeSettings));
             System.Windows.Forms.DataGridViewCellStyle dataGridViewCellStyle2 = new System.Windows.Forms.DataGridViewCellStyle();
+            System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(AdministrativeSettings));
             this.tabControl1 = new System.Windows.Forms.TabControl();
             this.tabPageModes = new System.Windows.Forms.TabPage();
             this.lblSelectModes = new System.Windows.Forms.Label();
@@ -85,6 +85,19 @@ namespace HearThis.UI
             this._btnAddRecordByVerseRange = new System.Windows.Forms.Button();
             this._verseCtrlRecordByVerseRangeStart = new SIL.Windows.Forms.Scripture.VerseControl();
             this._verseCtrlRecordByVerseRangeEnd = new SIL.Windows.Forms.Scripture.VerseControl();
+            this.tabPageClipEditor = new System.Windows.Forms.TabPage();
+            this._tableLayoutPanelClipEditor = new System.Windows.Forms.TableLayoutPanel();
+            this._lblInstructions = new System.Windows.Forms.Label();
+            this._rdoUseSpecifiedEditor = new System.Windows.Forms.RadioButton();
+            this._rdoUseDefaultAssociatedApplication = new System.Windows.Forms.RadioButton();
+            this._lblPathToWAVFileEditor = new System.Windows.Forms.Label();
+            this._lblWAVFileEditingApplicationName = new System.Windows.Forms.Label();
+            this._txtEditingApplicationName = new System.Windows.Forms.TextBox();
+            this._chkWAVEditorCommandLineArguments = new System.Windows.Forms.CheckBox();
+            this._txtCommandLineArguments = new System.Windows.Forms.TextBox();
+            this._btnOpenFileChooser = new System.Windows.Forms.Button();
+            this._lblCommandLineArgumentsInstructions = new System.Windows.Forms.Label();
+            this._lblWAVEditorCommandLineExample = new System.Windows.Forms.Label();
             this._btnOk = new System.Windows.Forms.Button();
             this.l10NSharpExtender1 = new L10NSharp.UI.L10NSharpExtender(this.components);
             _lblBreakClauses = new System.Windows.Forms.Label();
@@ -106,6 +119,8 @@ namespace HearThis.UI
             this.tabPageRecordByVerse.SuspendLayout();
             this._tableLayoutPanelRecordByVerse.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)(this._gridVerseRanges)).BeginInit();
+            this.tabPageClipEditor.SuspendLayout();
+            this._tableLayoutPanelClipEditor.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)(this.l10NSharpExtender1)).BeginInit();
             this.SuspendLayout();
             // 
@@ -167,9 +182,8 @@ namespace HearThis.UI
             _lblRecordByVerseRangeStart.AutoSize = true;
             this.l10NSharpExtender1.SetLocalizableToolTip(_lblRecordByVerseRangeStart, null);
             this.l10NSharpExtender1.SetLocalizationComment(_lblRecordByVerseRangeStart, null);
-            this.l10NSharpExtender1.SetLocalizationPriority(_lblRecordByVerseRangeStart, L10NSharp.LocalizationPriority.NotLocalizable);
             this.l10NSharpExtender1.SetLocalizingId(_lblRecordByVerseRangeStart, "AdministrativeSettings._lblRecordByVerseRangeStart");
-            _lblRecordByVerseRangeStart.Location = new System.Drawing.Point(106, 81);
+            _lblRecordByVerseRangeStart.Location = new System.Drawing.Point(112, 78);
             _lblRecordByVerseRangeStart.Name = "_lblRecordByVerseRangeStart";
             _lblRecordByVerseRangeStart.Size = new System.Drawing.Size(33, 13);
             _lblRecordByVerseRangeStart.TabIndex = 10;
@@ -181,9 +195,8 @@ namespace HearThis.UI
             _lblRecordByVerseRangeEnd.AutoSize = true;
             this.l10NSharpExtender1.SetLocalizableToolTip(_lblRecordByVerseRangeEnd, null);
             this.l10NSharpExtender1.SetLocalizationComment(_lblRecordByVerseRangeEnd, null);
-            this.l10NSharpExtender1.SetLocalizationPriority(_lblRecordByVerseRangeEnd, L10NSharp.LocalizationPriority.NotLocalizable);
             this.l10NSharpExtender1.SetLocalizingId(_lblRecordByVerseRangeEnd, "AdministrativeSettings._lblRecordByVerseRangeEnd");
-            _lblRecordByVerseRangeEnd.Location = new System.Drawing.Point(116, 111);
+            _lblRecordByVerseRangeEnd.Location = new System.Drawing.Point(122, 102);
             _lblRecordByVerseRangeEnd.Name = "_lblRecordByVerseRangeEnd";
             _lblRecordByVerseRangeEnd.Size = new System.Drawing.Size(23, 13);
             _lblRecordByVerseRangeEnd.TabIndex = 13;
@@ -199,6 +212,7 @@ namespace HearThis.UI
             this.tabControl1.Controls.Add(this.tabPagePunctuation);
             this.tabControl1.Controls.Add(this.tabPageInterface);
             this.tabControl1.Controls.Add(this.tabPageRecordByVerse);
+            this.tabControl1.Controls.Add(this.tabPageClipEditor);
             this.tabControl1.Location = new System.Drawing.Point(12, 12);
             this.tabControl1.Name = "tabControl1";
             this.tabControl1.SelectedIndex = 0;
@@ -218,6 +232,7 @@ namespace HearThis.UI
             this.tabPageModes.Padding = new System.Windows.Forms.Padding(3);
             this.tabPageModes.Size = new System.Drawing.Size(360, 343);
             this.tabPageModes.TabIndex = 0;
+            this.tabPageModes.Tag = "Modes";
             this.tabPageModes.Text = "Modes";
             // 
             // lblSelectModes
@@ -340,6 +355,7 @@ namespace HearThis.UI
             this.tabPageSkipping.Padding = new System.Windows.Forms.Padding(3);
             this.tabPageSkipping.Size = new System.Drawing.Size(360, 343);
             this.tabPageSkipping.TabIndex = 1;
+            this.tabPageSkipping.Tag = "Skipping";
             this.tabPageSkipping.Text = "Skipping";
             // 
             // _tableLayoutPanelSkipping
@@ -431,6 +447,7 @@ namespace HearThis.UI
             this.tabPagePunctuation.Name = "tabPagePunctuation";
             this.tabPagePunctuation.Size = new System.Drawing.Size(360, 343);
             this.tabPagePunctuation.TabIndex = 2;
+            this.tabPagePunctuation.Tag = "Punctuation";
             this.tabPagePunctuation.Text = "Punctuation";
             // 
             // _tableLayoutPanelPunctuation
@@ -652,6 +669,7 @@ namespace HearThis.UI
             this.tabPageInterface.Name = "tabPageInterface";
             this.tabPageInterface.Size = new System.Drawing.Size(360, 343);
             this.tabPageInterface.TabIndex = 3;
+            this.tabPageInterface.Tag = "Interface";
             this.tabPageInterface.Text = "Interface";
             // 
             // _groupAdvancedUI
@@ -686,6 +704,7 @@ namespace HearThis.UI
             this._tableLayoutPanelAdvancedUI.RowStyles.Add(new System.Windows.Forms.RowStyle());
             this._tableLayoutPanelAdvancedUI.RowStyles.Add(new System.Windows.Forms.RowStyle());
             this._tableLayoutPanelAdvancedUI.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 100F));
+            this._tableLayoutPanelAdvancedUI.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Absolute, 20F));
             this._tableLayoutPanelAdvancedUI.Size = new System.Drawing.Size(325, 125);
             this._tableLayoutPanelAdvancedUI.TabIndex = 13;
             // 
@@ -810,6 +829,7 @@ namespace HearThis.UI
             this.tabPageRecordByVerse.Padding = new System.Windows.Forms.Padding(3);
             this.tabPageRecordByVerse.Size = new System.Drawing.Size(360, 343);
             this.tabPageRecordByVerse.TabIndex = 4;
+            this.tabPageRecordByVerse.Tag = "RecordByVerse";
             this.tabPageRecordByVerse.Text = "Record by verse";
             this.tabPageRecordByVerse.UseVisualStyleBackColor = true;
             // 
@@ -848,7 +868,6 @@ namespace HearThis.UI
             this._lblRecordByVerse.Dock = System.Windows.Forms.DockStyle.Top;
             this.l10NSharpExtender1.SetLocalizableToolTip(this._lblRecordByVerse, null);
             this.l10NSharpExtender1.SetLocalizationComment(this._lblRecordByVerse, null);
-            this.l10NSharpExtender1.SetLocalizationPriority(this._lblRecordByVerse, L10NSharp.LocalizationPriority.NotLocalizable);
             this.l10NSharpExtender1.SetLocalizingId(this._lblRecordByVerse, "AdministrativeSettings.lblRecordByVerse");
             this._lblRecordByVerse.Location = new System.Drawing.Point(14, 11);
             this._lblRecordByVerse.Name = "_lblRecordByVerse";
@@ -891,7 +910,7 @@ namespace HearThis.UI
             this.l10NSharpExtender1.SetLocalizableToolTip(this._gridVerseRanges, null);
             this.l10NSharpExtender1.SetLocalizationComment(this._gridVerseRanges, null);
             this.l10NSharpExtender1.SetLocalizingId(this._gridVerseRanges, "AdministrativeSettings.gridVerseRanges");
-            this._gridVerseRanges.Location = new System.Drawing.Point(14, 175);
+            this._gridVerseRanges.Location = new System.Drawing.Point(14, 163);
             this._gridVerseRanges.MultiSelect = false;
             this._gridVerseRanges.Name = "_gridVerseRanges";
             this._gridVerseRanges.PaintFullRowFocusRectangle = true;
@@ -905,7 +924,7 @@ namespace HearThis.UI
             this._gridVerseRanges.SelectedRowForeColor = System.Drawing.Color.Empty;
             this._gridVerseRanges.SelectionMode = System.Windows.Forms.DataGridViewSelectionMode.FullRowSelect;
             this._gridVerseRanges.ShowWaterMarkWhenDirty = false;
-            this._gridVerseRanges.Size = new System.Drawing.Size(326, 96);
+            this._gridVerseRanges.Size = new System.Drawing.Size(326, 108);
             this._gridVerseRanges.TabIndex = 1;
             this._gridVerseRanges.TextBoxEditControlBorderColor = System.Drawing.Color.Silver;
             this._gridVerseRanges.WaterMark = "";
@@ -949,9 +968,8 @@ namespace HearThis.UI
             this._btnAddRecordByVerseRange.AutoSize = true;
             this.l10NSharpExtender1.SetLocalizableToolTip(this._btnAddRecordByVerseRange, null);
             this.l10NSharpExtender1.SetLocalizationComment(this._btnAddRecordByVerseRange, null);
-            this.l10NSharpExtender1.SetLocalizationPriority(this._btnAddRecordByVerseRange, L10NSharp.LocalizationPriority.NotLocalizable);
             this.l10NSharpExtender1.SetLocalizingId(this._btnAddRecordByVerseRange, "AdministrativeSettings._btnAddRecordByVerseRange");
-            this._btnAddRecordByVerseRange.Location = new System.Drawing.Point(265, 141);
+            this._btnAddRecordByVerseRange.Location = new System.Drawing.Point(265, 129);
             this._btnAddRecordByVerseRange.Margin = new System.Windows.Forms.Padding(3, 8, 3, 8);
             this._btnAddRecordByVerseRange.Name = "_btnAddRecordByVerseRange";
             this._btnAddRecordByVerseRange.Size = new System.Drawing.Size(75, 23);
@@ -974,7 +992,8 @@ namespace HearThis.UI
             this.l10NSharpExtender1.SetLocalizationComment(this._verseCtrlRecordByVerseRangeStart, null);
             this.l10NSharpExtender1.SetLocalizationPriority(this._verseCtrlRecordByVerseRangeStart, L10NSharp.LocalizationPriority.NotLocalizable);
             this.l10NSharpExtender1.SetLocalizingId(this._verseCtrlRecordByVerseRangeStart, "AdministrativeSettings.VerseControlStart");
-            this._verseCtrlRecordByVerseRangeStart.Location = new System.Drawing.Point(145, 76);
+            this._verseCtrlRecordByVerseRangeStart.Location = new System.Drawing.Point(148, 73);
+            this._verseCtrlRecordByVerseRangeStart.Margin = new System.Windows.Forms.Padding(0);
             this._verseCtrlRecordByVerseRangeStart.Name = "_verseCtrlRecordByVerseRangeStart";
             this._verseCtrlRecordByVerseRangeStart.Padding = new System.Windows.Forms.Padding(0, 0, 4, 0);
             this._verseCtrlRecordByVerseRangeStart.ShowEmptyBooks = true;
@@ -996,13 +1015,237 @@ namespace HearThis.UI
             this.l10NSharpExtender1.SetLocalizationComment(this._verseCtrlRecordByVerseRangeEnd, null);
             this.l10NSharpExtender1.SetLocalizationPriority(this._verseCtrlRecordByVerseRangeEnd, L10NSharp.LocalizationPriority.NotLocalizable);
             this.l10NSharpExtender1.SetLocalizingId(this._verseCtrlRecordByVerseRangeEnd, "AdministrativeSettings.VerseControl");
-            this._verseCtrlRecordByVerseRangeEnd.Location = new System.Drawing.Point(145, 106);
+            this._verseCtrlRecordByVerseRangeEnd.Location = new System.Drawing.Point(148, 97);
+            this._verseCtrlRecordByVerseRangeEnd.Margin = new System.Windows.Forms.Padding(0);
             this._verseCtrlRecordByVerseRangeEnd.Name = "_verseCtrlRecordByVerseRangeEnd";
             this._verseCtrlRecordByVerseRangeEnd.Padding = new System.Windows.Forms.Padding(0, 0, 4, 0);
             this._verseCtrlRecordByVerseRangeEnd.ShowEmptyBooks = true;
             this._verseCtrlRecordByVerseRangeEnd.Size = new System.Drawing.Size(195, 24);
             this._verseCtrlRecordByVerseRangeEnd.TabIndex = 15;
             this._verseCtrlRecordByVerseRangeEnd.Leave += new System.EventHandler(this.VerseCtrlRecordByVerseRangeChanged);
+            // 
+            // tabPageClipEditor
+            // 
+            this.tabPageClipEditor.Controls.Add(this._tableLayoutPanelClipEditor);
+            this.l10NSharpExtender1.SetLocalizableToolTip(this.tabPageClipEditor, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this.tabPageClipEditor, null);
+            this.l10NSharpExtender1.SetLocalizingId(this.tabPageClipEditor, "AdministrativeSettings.tabPageClipEditor");
+            this.tabPageClipEditor.Location = new System.Drawing.Point(4, 22);
+            this.tabPageClipEditor.Name = "tabPageClipEditor";
+            this.tabPageClipEditor.Padding = new System.Windows.Forms.Padding(3);
+            this.tabPageClipEditor.Size = new System.Drawing.Size(360, 343);
+            this.tabPageClipEditor.TabIndex = 5;
+            this.tabPageClipEditor.Tag = "ClipEditor";
+            this.tabPageClipEditor.Text = "Clip Editor";
+            this.tabPageClipEditor.UseVisualStyleBackColor = true;
+            // 
+            // _tableLayoutPanelClipEditor
+            // 
+            this._tableLayoutPanelClipEditor.BackColor = System.Drawing.SystemColors.ButtonFace;
+            this._tableLayoutPanelClipEditor.ColumnCount = 2;
+            this._tableLayoutPanelClipEditor.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle());
+            this._tableLayoutPanelClipEditor.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Percent, 100F));
+            this._tableLayoutPanelClipEditor.Controls.Add(this._lblInstructions, 0, 0);
+            this._tableLayoutPanelClipEditor.Controls.Add(this._rdoUseSpecifiedEditor, 0, 1);
+            this._tableLayoutPanelClipEditor.Controls.Add(this._rdoUseDefaultAssociatedApplication, 0, 7);
+            this._tableLayoutPanelClipEditor.Controls.Add(this._lblPathToWAVFileEditor, 0, 2);
+            this._tableLayoutPanelClipEditor.Controls.Add(this._lblWAVFileEditingApplicationName, 0, 3);
+            this._tableLayoutPanelClipEditor.Controls.Add(this._txtEditingApplicationName, 0, 4);
+            this._tableLayoutPanelClipEditor.Controls.Add(this._chkWAVEditorCommandLineArguments, 1, 3);
+            this._tableLayoutPanelClipEditor.Controls.Add(this._txtCommandLineArguments, 1, 4);
+            this._tableLayoutPanelClipEditor.Controls.Add(this._btnOpenFileChooser, 1, 1);
+            this._tableLayoutPanelClipEditor.Controls.Add(this._lblCommandLineArgumentsInstructions, 0, 5);
+            this._tableLayoutPanelClipEditor.Controls.Add(this._lblWAVEditorCommandLineExample, 1, 6);
+            this._tableLayoutPanelClipEditor.Dock = System.Windows.Forms.DockStyle.Fill;
+            this._tableLayoutPanelClipEditor.Location = new System.Drawing.Point(3, 3);
+            this._tableLayoutPanelClipEditor.Name = "_tableLayoutPanelClipEditor";
+            this._tableLayoutPanelClipEditor.Padding = new System.Windows.Forms.Padding(11);
+            this._tableLayoutPanelClipEditor.RowCount = 8;
+            this._tableLayoutPanelClipEditor.RowStyles.Add(new System.Windows.Forms.RowStyle());
+            this._tableLayoutPanelClipEditor.RowStyles.Add(new System.Windows.Forms.RowStyle());
+            this._tableLayoutPanelClipEditor.RowStyles.Add(new System.Windows.Forms.RowStyle());
+            this._tableLayoutPanelClipEditor.RowStyles.Add(new System.Windows.Forms.RowStyle());
+            this._tableLayoutPanelClipEditor.RowStyles.Add(new System.Windows.Forms.RowStyle());
+            this._tableLayoutPanelClipEditor.RowStyles.Add(new System.Windows.Forms.RowStyle());
+            this._tableLayoutPanelClipEditor.RowStyles.Add(new System.Windows.Forms.RowStyle());
+            this._tableLayoutPanelClipEditor.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 100F));
+            this._tableLayoutPanelClipEditor.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Absolute, 20F));
+            this._tableLayoutPanelClipEditor.Size = new System.Drawing.Size(354, 337);
+            this._tableLayoutPanelClipEditor.TabIndex = 7;
+            // 
+            // _lblInstructions
+            // 
+            this._lblInstructions.AutoSize = true;
+            this._tableLayoutPanelClipEditor.SetColumnSpan(this._lblInstructions, 2);
+            this.l10NSharpExtender1.SetLocalizableToolTip(this._lblInstructions, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this._lblInstructions, "Param 0: \"WAV\" (file format); Param 1: \"HearThis\" (program name); Param 2: \" Auda" +
+        "city\" (third-party product name); Param 3: Name of \"Check for Problems\" mode (as" +
+        " localized)");
+            this.l10NSharpExtender1.SetLocalizingId(this._lblInstructions, "AdministrativeSettings._lblInstructions");
+            this._lblInstructions.Location = new System.Drawing.Point(14, 11);
+            this._lblInstructions.Name = "_lblInstructions";
+            this._lblInstructions.Size = new System.Drawing.Size(321, 65);
+            this._lblInstructions.TabIndex = 6;
+            this._lblInstructions.Text = resources.GetString("_lblInstructions.Text");
+            // 
+            // _rdoUseSpecifiedEditor
+            // 
+            this._rdoUseSpecifiedEditor.AutoSize = true;
+            this._rdoUseSpecifiedEditor.Checked = true;
+            this.l10NSharpExtender1.SetLocalizableToolTip(this._rdoUseSpecifiedEditor, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this._rdoUseSpecifiedEditor, "Param 0: \"WAV\" (file format)");
+            this.l10NSharpExtender1.SetLocalizingId(this._rdoUseSpecifiedEditor, "AdministrativeSettings._rdoUseSpecifiedEditor");
+            this._rdoUseSpecifiedEditor.Location = new System.Drawing.Point(14, 84);
+            this._rdoUseSpecifiedEditor.Margin = new System.Windows.Forms.Padding(3, 8, 3, 3);
+            this._rdoUseSpecifiedEditor.Name = "_rdoUseSpecifiedEditor";
+            this._rdoUseSpecifiedEditor.Size = new System.Drawing.Size(151, 17);
+            this._rdoUseSpecifiedEditor.TabIndex = 8;
+            this._rdoUseSpecifiedEditor.TabStop = true;
+            this._rdoUseSpecifiedEditor.Text = "Use specified {0} file editor";
+            this._rdoUseSpecifiedEditor.UseVisualStyleBackColor = true;
+            this._rdoUseSpecifiedEditor.CheckedChanged += new System.EventHandler(this.HandleWavEditorOptionChanged);
+            // 
+            // _rdoUseDefaultAssociatedApplication
+            // 
+            this._rdoUseDefaultAssociatedApplication.AutoSize = true;
+            this.l10NSharpExtender1.SetLocalizableToolTip(this._rdoUseDefaultAssociatedApplication, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this._rdoUseDefaultAssociatedApplication, null);
+            this.l10NSharpExtender1.SetLocalizingId(this._rdoUseDefaultAssociatedApplication, "AdministrativeSettings._rdoUseDefaultAssociatedApplication");
+            this._rdoUseDefaultAssociatedApplication.Location = new System.Drawing.Point(14, 290);
+            this._rdoUseDefaultAssociatedApplication.Name = "_rdoUseDefaultAssociatedApplication";
+            this._rdoUseDefaultAssociatedApplication.Size = new System.Drawing.Size(133, 17);
+            this._rdoUseDefaultAssociatedApplication.TabIndex = 9;
+            this._rdoUseDefaultAssociatedApplication.Text = "Use default application";
+            this._rdoUseDefaultAssociatedApplication.UseVisualStyleBackColor = true;
+            this._rdoUseDefaultAssociatedApplication.CheckedChanged += new System.EventHandler(this.HandleWavEditorOptionChanged);
+            // 
+            // _lblPathToWAVFileEditor
+            // 
+            this._lblPathToWAVFileEditor.AutoEllipsis = true;
+            this._tableLayoutPanelClipEditor.SetColumnSpan(this._lblPathToWAVFileEditor, 2);
+            this._lblPathToWAVFileEditor.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.l10NSharpExtender1.SetLocalizableToolTip(this._lblPathToWAVFileEditor, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this._lblPathToWAVFileEditor, null);
+            this.l10NSharpExtender1.SetLocalizationPriority(this._lblPathToWAVFileEditor, L10NSharp.LocalizationPriority.NotLocalizable);
+            this.l10NSharpExtender1.SetLocalizingId(this._lblPathToWAVFileEditor, "AdministrativeSettings._lblPathToWAVFileEditor");
+            this._lblPathToWAVFileEditor.Location = new System.Drawing.Point(31, 104);
+            this._lblPathToWAVFileEditor.Margin = new System.Windows.Forms.Padding(20, 0, 8, 0);
+            this._lblPathToWAVFileEditor.Name = "_lblPathToWAVFileEditor";
+            this._lblPathToWAVFileEditor.Size = new System.Drawing.Size(304, 13);
+            this._lblPathToWAVFileEditor.TabIndex = 13;
+            this._lblPathToWAVFileEditor.Text = "#";
+            this._lblPathToWAVFileEditor.Visible = false;
+            // 
+            // _lblWAVFileEditingApplicationName
+            // 
+            this._lblWAVFileEditingApplicationName.Anchor = System.Windows.Forms.AnchorStyles.Left;
+            this._lblWAVFileEditingApplicationName.AutoSize = true;
+            this.l10NSharpExtender1.SetLocalizableToolTip(this._lblWAVFileEditingApplicationName, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this._lblWAVFileEditingApplicationName, null);
+            this.l10NSharpExtender1.SetLocalizingId(this._lblWAVFileEditingApplicationName, "AdministrativeSettings._lblWAVFileEditingApplicationName");
+            this._lblWAVFileEditingApplicationName.Location = new System.Drawing.Point(31, 125);
+            this._lblWAVFileEditingApplicationName.Margin = new System.Windows.Forms.Padding(20, 6, 3, 4);
+            this._lblWAVFileEditingApplicationName.Name = "_lblWAVFileEditingApplicationName";
+            this._lblWAVFileEditingApplicationName.Size = new System.Drawing.Size(125, 13);
+            this._lblWAVFileEditingApplicationName.TabIndex = 15;
+            this._lblWAVFileEditingApplicationName.Text = "Editing application name:";
+            this._lblWAVFileEditingApplicationName.Visible = false;
+            // 
+            // _txtEditingApplicationName
+            // 
+            this._txtEditingApplicationName.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+            this._txtEditingApplicationName.Cursor = System.Windows.Forms.Cursors.Default;
+            this.l10NSharpExtender1.SetLocalizableToolTip(this._txtEditingApplicationName, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this._txtEditingApplicationName, null);
+            this.l10NSharpExtender1.SetLocalizationPriority(this._txtEditingApplicationName, L10NSharp.LocalizationPriority.NotLocalizable);
+            this.l10NSharpExtender1.SetLocalizingId(this._txtEditingApplicationName, "AdministrativeSettings._txtEditingApplicationName");
+            this._txtEditingApplicationName.Location = new System.Drawing.Point(31, 148);
+            this._txtEditingApplicationName.Margin = new System.Windows.Forms.Padding(20, 3, 3, 3);
+            this._txtEditingApplicationName.MinimumSize = new System.Drawing.Size(100, 20);
+            this._txtEditingApplicationName.Name = "_txtEditingApplicationName";
+            this._txtEditingApplicationName.Size = new System.Drawing.Size(134, 20);
+            this._txtEditingApplicationName.TabIndex = 16;
+            this._txtEditingApplicationName.Visible = false;
+            // 
+            // _chkWAVEditorCommandLineArguments
+            // 
+            this._chkWAVEditorCommandLineArguments.Anchor = System.Windows.Forms.AnchorStyles.Left;
+            this._chkWAVEditorCommandLineArguments.AutoSize = true;
+            this._chkWAVEditorCommandLineArguments.Enabled = false;
+            this.l10NSharpExtender1.SetLocalizableToolTip(this._chkWAVEditorCommandLineArguments, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this._chkWAVEditorCommandLineArguments, null);
+            this.l10NSharpExtender1.SetLocalizingId(this._chkWAVEditorCommandLineArguments, "AdministrativeSettings._chkWAVEditorCommandLineArguments");
+            this._chkWAVEditorCommandLineArguments.Location = new System.Drawing.Point(171, 125);
+            this._chkWAVEditorCommandLineArguments.Margin = new System.Windows.Forms.Padding(3, 8, 3, 3);
+            this._chkWAVEditorCommandLineArguments.Name = "_chkWAVEditorCommandLineArguments";
+            this._chkWAVEditorCommandLineArguments.Size = new System.Drawing.Size(169, 17);
+            this._chkWAVEditorCommandLineArguments.TabIndex = 12;
+            this._chkWAVEditorCommandLineArguments.Text = "Command-line arguments required";
+            this._chkWAVEditorCommandLineArguments.UseVisualStyleBackColor = true;
+            this._chkWAVEditorCommandLineArguments.Visible = false;
+            this._chkWAVEditorCommandLineArguments.CheckedChanged += new System.EventHandler(this._chkWAVEditorCommandLineArguments_CheckedChanged);
+            // 
+            // _txtCommandLineArguments
+            // 
+            this._txtCommandLineArguments.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+            this.l10NSharpExtender1.SetLocalizableToolTip(this._txtCommandLineArguments, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this._txtCommandLineArguments, null);
+            this.l10NSharpExtender1.SetLocalizationPriority(this._txtCommandLineArguments, L10NSharp.LocalizationPriority.NotLocalizable);
+            this.l10NSharpExtender1.SetLocalizingId(this._txtCommandLineArguments, "AdministrativeSettings._txtCommandLineArguments");
+            this._txtCommandLineArguments.Location = new System.Drawing.Point(171, 148);
+            this._txtCommandLineArguments.Name = "_txtCommandLineArguments";
+            this._txtCommandLineArguments.Size = new System.Drawing.Size(169, 20);
+            this._txtCommandLineArguments.TabIndex = 11;
+            this._txtCommandLineArguments.Visible = false;
+            // 
+            // _btnOpenFileChooser
+            // 
+            this._btnOpenFileChooser.Anchor = System.Windows.Forms.AnchorStyles.Left;
+            this._btnOpenFileChooser.Image = global::HearThis.Properties.Resources.ellipsis;
+            this.l10NSharpExtender1.SetLocalizableToolTip(this._btnOpenFileChooser, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this._btnOpenFileChooser, null);
+            this.l10NSharpExtender1.SetLocalizationPriority(this._btnOpenFileChooser, L10NSharp.LocalizationPriority.NotLocalizable);
+            this.l10NSharpExtender1.SetLocalizingId(this._btnOpenFileChooser, "AdministrativeSettings._btnOpenFileChooser");
+            this._btnOpenFileChooser.Location = new System.Drawing.Point(171, 80);
+            this._btnOpenFileChooser.Name = "_btnOpenFileChooser";
+            this._btnOpenFileChooser.Size = new System.Drawing.Size(24, 20);
+            this._btnOpenFileChooser.TabIndex = 14;
+            this._btnOpenFileChooser.UseVisualStyleBackColor = true;
+            this._btnOpenFileChooser.Click += new System.EventHandler(this._btnOpenFileChooser_Click);
+            // 
+            // _lblCommandLineArgumentsInstructions
+            // 
+            this._lblCommandLineArgumentsInstructions.AutoSize = true;
+            this._tableLayoutPanelClipEditor.SetColumnSpan(this._lblCommandLineArgumentsInstructions, 2);
+            this.l10NSharpExtender1.SetLocalizableToolTip(this._lblCommandLineArgumentsInstructions, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this._lblCommandLineArgumentsInstructions, "Param 0: \"HearThis\" (product name); Param 1: \"{path}\" (literal text)");
+            this.l10NSharpExtender1.SetLocalizingId(this._lblCommandLineArgumentsInstructions, "AdministrativeSettings._lblCommandLineArgumentsInstructions");
+            this._lblCommandLineArgumentsInstructions.Location = new System.Drawing.Point(31, 177);
+            this._lblCommandLineArgumentsInstructions.Margin = new System.Windows.Forms.Padding(20, 6, 3, 0);
+            this._lblCommandLineArgumentsInstructions.Name = "_lblCommandLineArgumentsInstructions";
+            this._lblCommandLineArgumentsInstructions.Size = new System.Drawing.Size(309, 91);
+            this._lblCommandLineArgumentsInstructions.TabIndex = 17;
+            this._lblCommandLineArgumentsInstructions.Text = resources.GetString("_lblCommandLineArgumentsInstructions.Text");
+            this._lblCommandLineArgumentsInstructions.Visible = false;
+            // 
+            // _lblWAVEditorCommandLineExample
+            // 
+            this._lblWAVEditorCommandLineExample.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
+            this._lblWAVEditorCommandLineExample.AutoSize = true;
+            this._lblWAVEditorCommandLineExample.Font = new System.Drawing.Font("Microsoft Sans Serif", 8.25F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.l10NSharpExtender1.SetLocalizableToolTip(this._lblWAVEditorCommandLineExample, null);
+            this.l10NSharpExtender1.SetLocalizationComment(this._lblWAVEditorCommandLineExample, "Param  is \"{path}\" (literal text)");
+            this.l10NSharpExtender1.SetLocalizationPriority(this._lblWAVEditorCommandLineExample, L10NSharp.LocalizationPriority.NotLocalizable);
+            this.l10NSharpExtender1.SetLocalizingId(this._lblWAVEditorCommandLineExample, "AdministrativeSettings._lblWAVEditorCommandLineExample");
+            this._lblWAVEditorCommandLineExample.Location = new System.Drawing.Point(282, 274);
+            this._lblWAVEditorCommandLineExample.Margin = new System.Windows.Forms.Padding(0, 6, 3, 0);
+            this._lblWAVEditorCommandLineExample.Name = "_lblWAVEditorCommandLineExample";
+            this._lblWAVEditorCommandLineExample.Size = new System.Drawing.Size(58, 13);
+            this._lblWAVEditorCommandLineExample.TabIndex = 18;
+            this._lblWAVEditorCommandLineExample.Text = "-i {0} -o {0}";
+            this._lblWAVEditorCommandLineExample.Visible = false;
             // 
             // _btnOk
             // 
@@ -1066,6 +1309,9 @@ namespace HearThis.UI
             this._tableLayoutPanelRecordByVerse.ResumeLayout(false);
             this._tableLayoutPanelRecordByVerse.PerformLayout();
             ((System.ComponentModel.ISupportInitialize)(this._gridVerseRanges)).EndInit();
+            this.tabPageClipEditor.ResumeLayout(false);
+            this._tableLayoutPanelClipEditor.ResumeLayout(false);
+            this._tableLayoutPanelClipEditor.PerformLayout();
             ((System.ComponentModel.ISupportInitialize)(this.l10NSharpExtender1)).EndInit();
             this.ResumeLayout(false);
 
@@ -1122,5 +1368,18 @@ namespace HearThis.UI
 		private System.Windows.Forms.DataGridViewTextBoxColumn colTo;
 		private SIL.Windows.Forms.Scripture.VerseControl _verseCtrlRecordByVerseRangeStart;
 		private SIL.Windows.Forms.Scripture.VerseControl _verseCtrlRecordByVerseRangeEnd;
+		private System.Windows.Forms.TabPage tabPageClipEditor;
+		private System.Windows.Forms.TableLayoutPanel _tableLayoutPanelClipEditor;
+		private System.Windows.Forms.Label _lblInstructions;
+		private System.Windows.Forms.RadioButton _rdoUseSpecifiedEditor;
+		private System.Windows.Forms.RadioButton _rdoUseDefaultAssociatedApplication;
+		private System.Windows.Forms.TextBox _txtCommandLineArguments;
+		private System.Windows.Forms.CheckBox _chkWAVEditorCommandLineArguments;
+		private System.Windows.Forms.Button _btnOpenFileChooser;
+		private System.Windows.Forms.TextBox _txtEditingApplicationName;
+		private System.Windows.Forms.Label _lblWAVFileEditingApplicationName;
+		private System.Windows.Forms.Label _lblCommandLineArgumentsInstructions;
+		private System.Windows.Forms.Label _lblWAVEditorCommandLineExample;
+		private System.Windows.Forms.Label _lblPathToWAVFileEditor;
 	}
 }